Move pattern.utils to helper
