+++ xawscf/pattern/apigateway.py
@@ -1,6 +1,6 @@
 import yaml
 from ..utils.yaml import Custom
-from .utils import get_full_name, make_output
+from ..utils.helper import get_full_name, make_output
 from .base_resource import BaseResource
 from .role import Role
