+++ xawscf/pattern/bucket.py
@@ -1,5 +1,5 @@
 from ..utils.yaml import Custom
-from .utils import set_tags_list, make_output
+from ..utils.helper import set_tags_list, make_output
 from .base_resource import BaseResource
 
 class Bucket(BaseResource):
