+++ xawscf/pattern/dynamodb_table.py
@@ -1,5 +1,5 @@
 from ..utils.yaml import Custom
-from .utils import try_set_field, make_output, set_tags_list
+from ..utils.helper import try_set_field, make_output, set_tags_list
 from .base import Base
 from .base_resource import BaseResource
 from .role import Role
