+++ xawscf/pattern/statemachine.py
@@ -1,6 +1,6 @@
 import json
 from ..utils.yaml import Custom
-from .utils import get_full_name, make_output, try_set_field
+from ..utils.helper import get_full_name, make_output, try_set_field
 from .base_resource import BaseResource
 from .role import Role
